Move detect prop position to HuskyVision and delete Claw
